backend/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
plaid-python==3.4.0
PyMySQL==0.9.3
requests==2.22.0
urllib3==1.25.8